+++ lib/cocoapods-core/podfile/dsl.rb
@@ -314,14 +314,14 @@ module Pod
       # Specifies the platform for which a static library should be built.
       #
       # CocoaPods provides a default deployment target if one is not specified.
-      # The current default values are `4.3` for iOS and `10.6` for OS X.
+      # The current default values are `4.3` for iOS, `10.6` for OS X and `2.0` for watchOS.
       #
       # If the deployment target requires it (iOS < `4.3`), `armv6`
       # architecture will be added to `ARCHS`.
       #
       # @param    [Symbol] name
-      #           the name of platform, can be either `:osx` for OS X or `:ios`
-      #           for iOS.
+      #           the name of platform, can be either `:osx` for OS X, `:ios`
+      #           for iOS or `:watchos` for watchOS.
       #
       # @param    [String, Version] target
       #           The optional deployment.  If not provided a default value