+++ spec/platform_spec.rb
@@ -6,6 +6,7 @@ module Pod
       it 'returns a new Platform instance' do
         Platform.ios.should == Platform.new(:ios)
         Platform.osx.should == Platform.new(:osx)
+        Platform.watchos.should == Platform.new(:watchos)
       end
 
       it 'can be initialized from another platform' do
@@ -43,8 +44,10 @@ module Pod
       it 'presents an accurate string representation' do
         @platform.to_s.should == 'iOS'
         Platform.new(:osx).to_s.should == 'OS X'
+        Platform.new(:watchos).to_s.should == 'watchOS'
         Platform.new(:ios, '5.0.0').to_s.should == 'iOS 5.0.0'
         Platform.new(:osx, '10.7').to_s.should  == 'OS X 10.7'
+        Platform.new(:watchos, '2.0').to_s.should == 'watchOS 2.0'
       end
 
       it "uses it's name as it's symbold version" do
@@ -80,6 +83,7 @@ module Pod
       it 'returns whether it requires legacy iOS architectures' do
         Platform.new(:ios, '4.0').requires_legacy_ios_archs?.should.be.true
         Platform.new(:ios, '5.0').requires_legacy_ios_archs?.should.be.false
+        Platform.new(:watchos, '2.0').requires_legacy_ios_archs?.should.be.false
       end
 
       it 'is usable as hash keys' do
@@ -106,6 +110,11 @@ module Pod
           Platform.new(:ios, '7.0').should.not.supports_dynamic_frameworks
           Platform.new(:ios, '8.0').should.supports_dynamic_frameworks
           Platform.new(:ios, '8.1').should.supports_dynamic_frameworks
+        end
+
+        it 'supports dynamic frameworks on watchOS' do
+          Platform.watchos.should.supports_dynamic_frameworks
+          Platform.new(:watchos, '2.0').should.supports_dynamic_frameworks
         end
       end
     end